chore: move misplaed logger.ts file
It ended up outside src by mistake.
