+++ apps/mana/src/stark/logger.ts
@@ -1,4 +1,4 @@
-import logger from '../../logger';
+import logger from '../logger';
 
 export default logger.child({
   module: 'stark'
